Change default w value
Algorithm is sensitive to values of w, especially in 3D. Lower values generally
result in slower convergence, higher values sometimes fail to converge. 0.8 was
found not to work with some 3D tests, so 0.7 seems a better default. 2D is
generally fine up to 0.9. Extensive characterization may be useful in future.