Add pre-normalization
The initial guesses weren't enough to ensure convergence on some data sets.
This adds pre-normalization. The normalization is then folded into the final
translation, rotation and scale output, so the function is the same. Due to
this normalization occuring, an initial guess for translation and scale is no
longer taken.

Unlike the paper, the span abs(X).max() is used. This deals better with
duplicated points than the variance. A 90th percentile or something might be
even more resistant to outliers.