Remove pre-normalization, use standard deviation
The pre-normalization was an extra step, and would have required handling
differently with affine/drift transforms.

Second, use the standard deviation (i.e. normalize to same variance) for the
scale estimate. This in some cases reduces the range of 'w' over which an
alignment converges. 'w' generally needs to be smaller than before, as the
standard deviation is smaller than the span.