Change axis of RMSD
X is the template, and Y is the set to be registered in other functions.

The points in the aligned set should cover the template set as best possible
(not vice versa). This prevents a false minimum being found where all aligned
points are collapsed onto a single template point.